refactor: REST client for HttpClient injection support
Introduce HttpClientBackedRestClient to enable use of externally managed HttpClient instances with custom configuration.

Refactor RestClientBase to be abstract and move generic request logic there. Update RestClient to handle HttpClient-specific features. Expand ICustomRestClient interface with additional properties and events.

Update service client constructors to accept Uri, ICustomRestClient, or HttpClient. Replace RestClientBase usages in context classes with RestClient or ICustomRestClient.

These changes improve extensibility, separation of concerns, and support advanced scenarios such as custom authentication and handler chains.
